Revert collection back to storing tokens
- it's kind of nice to be able to have any type
of state in a collection. And if we return a
state reference instead of a token from the at
method, then we have problems where someone might
want to use the meta token. So it's best to just
return a token so we get all that functionality
even if it means we create and store tokens for
all the items in the collection.